Uploading test code for software serial communication.
This commit consists of several differernt software serial test codes.
I used a new SoftwareSerial library, AltSoftSerial, instead of the
built-in library because I heard that the built-in library has high
latency which might be the reason why the code wasn't working in the
first place since the delay caused the DMX library to fault. These are
only prototypes and are not fully tested yet. If the code works it will
be merged to the main program.